src/requirements.txt
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.1.0
face-recognition==1.3.0
schedule==1.2.0
cryptography==41.0.7
pyjwt==2.8.0
dlib==19.24.2
numpy==1.24.3
werkzeug==3.0.1
